k8s-apps/backend-api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
confluent-kafka==2.4.0
mysql-connector-python==8.4.0
